Create README.md
# Discord-Bot
Discord Bot using Python

Firstly you need to create a discord account and create a server. 
Now head over to discord developer portal and simpling heading over to applications to make a new application.
It will automatically connect the bot with your server or the server you have admin previlages.
A bot with the permissions you want to give. Becareful while giving the permisions, the bot can even delete the channel with admin previlages. 
You would get a token in the Bot section of the selected Application.
Remeber to turn on MESSAGE CONTENT INTENT, used for the bot to receive the messages.
Now make a python project and using local host terminal of powershell install discord library and paste all the files.
Add your own token in code.
PS. The one i used will not work because it was my bot's token.

Simply run the program and it will connect to discord live server and the bot will be online and will be ready to use.

Enjoy!!

You can add more responses in the responses.py file by adding more if conditions.